--- grupo6/src/Pages/Home/Home.css
@@ -1,60 +1,0 @@
-@font-face {
-    font-family: Roboto;
-    src: url(../../../public/fonte/Roboto/Roboto-Bold.ttf);
-}
-
-
-
-.textos{
-    font-family: Roboto;
-    font-size: 20px;
-}
-
-.Head{
-    display: flex;
-    background-color: #005CB4;
-    padding-left: 10%;
-    padding-right: 10%;
-    height: auto;
-    min-height: 50px;
-    align-items: center;
-    flex-direction: row;
-    width: 100%;
-}
-
-.HeadItem{
-    display: flex;
-    flex-direction: row;
-    width: 70px;
-    
-}
-.grow1{
-    flex-grow: 1;
-}
-
-.grow2{
-    flex-grow: 5;
-}
-
-.MenuBar{
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-    flex-grow: 1;
-    flex-basis: auto;
-    width: 100px;
-}
-
-.MenuTool{
-    color: white;
-    align-items: center;
-    width: auto;
-    flex-grow: 1;
-}
-
-.Login{
-    display: grid;
-    grid-template-columns: repeat(auto-fit,minmax(40px,auto));
-    max-width: 250px;
-}
