+++ grupo6/src/Utilitarios/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Head";
